Add RouteLineLayerPositionTests
